+++ packages/package.go
@@ -60,12 +60,10 @@ type Package struct {
 	License         string  `config:"license,omitempty" json:"license,omitempty" yaml:"license,omitempty"`
 	versionSemVer   *semver.Version
 	Categories      []string         `config:"categories" json:"categories"`
-	Conditions      *Conditions      `config:"conditions,omitempty" json:"conditions,omitempty" yaml:"conditions,omitempty"`
 	Screenshots     []Image          `config:"screenshots,omitempty" json:"screenshots,omitempty" yaml:"screenshots,omitempty"`
 	Assets          []string         `config:"assets,omitempty" json:"assets,omitempty" yaml:"assets,omitempty"`
 	PolicyTemplates []PolicyTemplate `config:"policy_templates,omitempty" json:"policy_templates,omitempty" yaml:"policy_templates,omitempty"`
 	DataStreams     []*DataStream    `config:"data_streams,omitempty" json:"data_streams,omitempty" yaml:"data_streams,omitempty"`
-	Owner           *Owner           `config:"owner,omitempty" json:"owner,omitempty" yaml:"owner,omitempty"`
 	Vars            []Variable       `config:"vars" json:"vars,omitempty" yaml:"vars,omitempty"`
 
 	// Local path to the package dir
@@ -89,14 +87,18 @@ type BasePackage struct {
 	Icons               []Image              `config:"icons,omitempty" json:"icons,omitempty" yaml:"icons,omitempty"`
 	Internal            bool                 `config:"internal,omitempty" json:"internal,omitempty" yaml:"internal,omitempty"`
 	BasePolicyTemplates []BasePolicyTemplate `json:"policy_templates,omitempty"`
+	Conditions          *Conditions          `config:"conditions,omitempty" json:"conditions,omitempty" yaml:"conditions,omitempty"`
+	Owner               *Owner               `config:"owner,omitempty" json:"owner,omitempty" yaml:"owner,omitempty"`
+	Categories          []string             `config:"categories,omitempty" json:"categories,omitempty" yaml:"categories,omitempty"`
 }
 
 // BasePolicyTemplate is used for the package policy templates in the /search endpoint
 type BasePolicyTemplate struct {
-	Name        string  `config:"name" json:"name" validate:"required"`
-	Title       string  `config:"title" json:"title" validate:"required"`
-	Description string  `config:"description" json:"description" validate:"required"`
-	Icons       []Image `config:"icons,omitempty" json:"icons,omitempty" yaml:"icons,omitempty"`
+	Name        string   `config:"name" json:"name" validate:"required"`
+	Title       string   `config:"title" json:"title" validate:"required"`
+	Description string   `config:"description" json:"description" validate:"required"`
+	Icons       []Image  `config:"icons,omitempty" json:"icons,omitempty" yaml:"icons,omitempty"`
+	Categories  []string `config:"categories,omitempty" json:"categories,omitempty" yaml:"categories,omitempty"`
 }
 
 type PolicyTemplate struct {
@@ -113,8 +115,13 @@ type PolicyTemplate struct {
 }
 
 type Conditions struct {
-	KibanaVersion    string `config:"kibana.version,omitempty" json:"kibana.version,omitempty" yaml:"kibana.version,omitempty"`
-	kibanaConstraint *semver.Constraints
+	Kibana *KibanaConditions `config:"kibana,omitempty" json:"kibana,omitempty" yaml:"kibana,omitempty"`
+}
+
+// KibanaConditions defines conditions for Kibana (e.g. required version).
+type KibanaConditions struct {
+	Version    string `config:"version" json:"version" yaml:"version"`
+	constraint *semver.Constraints
 }
 
 type Version struct {
@@ -197,6 +204,7 @@ func NewPackage(basePath string, fsBuilder FileSystemBuilder) (*Package, error) 
 			Name:        t.Name,
 			Title:       t.Title,
 			Description: t.Description,
+			Categories:  t.Categories,
 		}
 
 		for k, i := range p.PolicyTemplates[i].Icons {
@@ -255,10 +263,10 @@ func NewPackage(basePath string, fsBuilder FileSystemBuilder) (*Package, error) 
 		}
 	}
 
-	if p.Conditions != nil && p.Conditions.KibanaVersion != "" {
-		p.Conditions.kibanaConstraint, err = semver.NewConstraint(p.Conditions.KibanaVersion)
+	if p.Conditions != nil && p.Conditions.Kibana != nil {
+		p.Conditions.Kibana.constraint, err = semver.NewConstraint(p.Conditions.Kibana.Version)
 		if err != nil {
-			return nil, errors.Wrapf(err, "invalid Kibana versions range: %s", p.Conditions.KibanaVersion)
+			return nil, errors.Wrapf(err, "invalid Kibana versions range: %s", p.Conditions.Kibana.Version)
 		}
 	}
 
@@ -315,13 +323,12 @@ func (p *Package) HasPolicyTemplateWithCategory(category string) bool {
 }
 
 func (p *Package) HasKibanaVersion(version *semver.Version) bool {
-
 	// If the version is not specified, it is for all versions
-	if p.Conditions == nil || version == nil || p.Conditions.kibanaConstraint == nil {
+	if p.Conditions == nil || p.Conditions.Kibana == nil || p.Conditions.Kibana.constraint == nil || version == nil {
 		return true
 	}
 
-	return p.Conditions.kibanaConstraint.Check(version)
+	return p.Conditions.Kibana.constraint.Check(version)
 }
 
 func (p *Package) IsNewerOrEqual(pp *Package) bool {