+++ packages/packages.go
@@ -28,10 +28,10 @@ type Packages []*Package
 func (p Packages) Len() int      { return len(p) }
 func (p Packages) Swap(i, j int) { p[i], p[j] = p[j], p[i] }
 func (p Packages) Less(i, j int) bool {
-	if p[i].Name == p[j].Name {
-		return p[i].Version < p[j].Version
+	if p[i].Title != nil && p[j].Title != nil && *p[i].Title != *p[j].Title {
+		return *p[i].Title < *p[j].Title
 	}
-	return p[i].Name < p[j].Name
+	return p[i].Version < p[j].Version
 }
 
 // Join returns a set of packages that combines both sets.